polish db pool hardening: fix bug, deduplicate, tighten validation
- Fix unhandled rejection in pool.on('connect'): the SET query's
  promise was not caught, so a failure would crash the process
- Add .int() to all pool env vars to reject fractional values
- Use consistent error shape (instanceof Error guard) in pool.on('error')
- Deduplicate requireDb() — now a const alias for getDb()
- Remove duplicate tutorialSteps query in initializeDatabase()

+++ server/config/env.ts
@@ -126,10 +126,10 @@ const envSchema = z.object({
   CHECK_IN_RADIUS_METERS: z.coerce.number().positive().max(150, "CHECK_IN_RADIUS_METERS cannot exceed 150m (service hard cap)").default(100),
 
   // Database pool tuning (all optional with safe defaults)
-  DB_POOL_MAX: z.coerce.number().positive().default(20),
-  DB_POOL_IDLE_TIMEOUT_MS: z.coerce.number().nonnegative().default(30000),
-  DB_POOL_CONNECTION_TIMEOUT_MS: z.coerce.number().positive().default(5000),
-  DB_STATEMENT_TIMEOUT_MS: z.coerce.number().positive().default(30000),
+  DB_POOL_MAX: z.coerce.number().int().positive().default(20),
+  DB_POOL_IDLE_TIMEOUT_MS: z.coerce.number().int().nonnegative().default(30000),
+  DB_POOL_CONNECTION_TIMEOUT_MS: z.coerce.number().int().positive().default(5000),
+  DB_STATEMENT_TIMEOUT_MS: z.coerce.number().int().positive().default(30000),
 });
 
 function validateEnv() {
